Proyecto/Redis/requirements.txt
flask
simplejson
redis
flask-socketio
mysql-connector-python
